Rename shortcuts.py file to shortcutssummary.py
