+++ spyder/app/mainwindow.py
@@ -2628,7 +2628,7 @@ class MainWindow(QMainWindow):
 
     @Slot()
     def show_shortcuts_dialog(self):
-        from spyder.widgets.shortcuts import ShortCutsSummaryDialog
+        from spyder.widgets.shortcutssummary import ShortCutsSummaryDialog
         dlg = ShortCutsSummaryDialog(None)
         dlg.exec_()
