+++ spyder/widgets/shortcutssummary.py
